Model Transaksi Tampilan Kepala Dinas done

+++ app/Http/Controllers/TransaksiController.php
@@ -61,8 +61,8 @@ class TransaksiController extends Controller
             'user_name' => 'required',
             'jenis_sampah_id' => 'required|exists:jenis_sampahs,id',
             'berat_kg' => 'required|numeric',
-            'catatan_verifikasi' => 'nullable|string',
-        ]);
+           'catatan_verifikasi' => 'nullable|string'
+       ]);
 
         if ($validator->fails()) {
             return redirect()->back()->withErrors($validator)->withInput();
@@ -125,18 +125,21 @@ class TransaksiController extends Controller
      */
     public function update(Request $request, Transaksi $transaksi)
     {
+         if (Auth::user()->role == 'kepala_dinas') {
+            return redirect()->route('transaksi.index')->with('error', 'Kepala Dinas tidak memiliki akses untuk mengubah transaksi.');
+        }
         $validator = Validator::make($request->all(), [
             'penjemputan_id' => 'nullable|exists:penjemputan,id',
             'user_id' => 'required|exists:users,id',
             'jenis_sampah_id' => 'required|exists:jenis_sampahs,id',
             'berat_kg' => 'required|numeric',
-            'harga_per_kilo_saat_transaksi' => 'required|numeric',
-            'nilai_saldo' => 'required|numeric',
-            'tanggal_transaksi' => 'required|date',
-            'dicatat_oleh_user_id' => 'nullable|exists:users,id',
-            'status_verifikasi' => 'nullable|in:terverifikasi,ditolak',
-            'catatan_verifikasi' => 'nullable|string',
-        ]);
+           'harga_per_kilo_saat_transaksi' => 'required|numeric',
+           'nilai_saldo' => 'required|numeric',
+           'tanggal_transaksi' => 'required|date',
+           'dicatat_oleh_user_id' => 'nullable|exists:users,id',
+           'status_verifikasi' => 'nullable|in:terverifikasi,ditolak',
+           'catatan_verifikasi' => 'nullable|string',
+       ]);
 
         if ($validator->fails()) {
             return redirect()->back()->withErrors($validator)->withInput();
@@ -154,6 +157,9 @@ class TransaksiController extends Controller
      */
     public function destroy(Transaksi $transaksi)
     {
+        if (Auth::user()->role == 'kepala_dinas') {
+            return redirect()->route('transaksi.index')->with('error', 'Kepala Dinas tidak memiliki akses untuk menghapus transaksi.');
+        }
         $transaksi->delete();
 
         return redirect()->route('transaksi.index')->with('success', 'Transaksi berhasil dihapus');
